main/postmarketos-ui-phosh: Modernize description and URL (MR 4366)
While a lot of Phosh development still happens on and for the Librem 5
we (thankfully) have lots of users and contributors using other devices.
Update the description to reflect that.

Ideally we'd emphasize that this is GNOME technology based and uses
wlroots but let's not  make the description too long.

Instead of Purism's generic home page point the URL to phosh.mobi which
has some more details. An alternative URL would be
https://puri.sm/pureos/phosh/ but that points to a page that is a
currently a bit dated.

[ci:skip-vercheck]
[ci:skip-build] already built successfully in CI