device-nokia-n900: Add Portugese/Spanish keymap (MR 4379)
Add the Portugese/Spanish keymap. Similarly to the Finnish/Swedish
keymap, it's compatible with both Portugese and Spanish keyboards.

Also, modify x11-keymap so that the Portugese/Spanish map will use
Alt+Shift to enter an underscore in the F key and a backslash in the C
key, which was done to avoid redundancy with the "Ñ" key.

Signed-off-by: Lux Aliaga <lux@nixgoat.me

[ci:skip-build] already built successfully in CI