device-nvidia-tegra-armv7: introduce common armv7 Tegra device package (MR 4606)
This device port uses U-Boot as a basic bootloader. Users should install
U-Boot and reinstall pmOS to the U-Boot version of port (this one).
Thus an upgrade from old installations that depend on old vendor bootloader
is not implemented.

Instruction how to install U-Boot can be found on dedicated page of
pmOS wiki.

Signed-off-by: Svyatoslav Ryhel <clamor95@gmail.com>