device-generic-x86_64: new aport (MR 4554)
This device package is intended to replace the tablet-x*uefi device
packages. It was made by basically merging config from those two
packages.

One important change (besides the name, covered later) is that it
replaces grub-efi with sd-boot for these devices.

The pkgver is starting at '8' to get ahead of the packages this is meant
to replace.

Because package names, especially for device packages(!), is important,
here's a long explanation of why the name is "device-generic-x86_64":

"uefi" was excluded from the device package name, because it should (in
theory at least...) be able to simultaneously support EFI and non-EFI
boot in 1 device package. The device-qemu-amd64 package kiiiinda does
this already and could be a useful example for how to do that, though in
this situation we'd need to support syslinux for legacy boot instead of
what qemu does (bypass installed bootloader and directly boot the
kernel). In other words, the packages this one is meant to replace only
support EFI boot, they currently don't support legacy boot. So IF we
need to support legacy boot in the future and IF there's no way to
support both in one package, THEN we can consider changing the name of
this one or calling the other "-legacy" or something.

"x86_64" instead of "amd64" (like the device-qemu-amd64 device package),
because "x86_64" is what Alpine uses to refer to this arch, and it
doesn't have a hw vendor name baked into it.

"tablet" was dropped from the name, because there are many other x86_64 devices
out there that people might want to install pmOS. For example, it might be
confusing if you want to install pmOS on an x86_64 toaster but best
pmOS image for it is named "device-tablet-..."