xiaomi-clover: add Mi Pad 4 Plus variation (MR 4702)
Similarly to how it was done for xiaomi-beyllium with varying
display panels, add kernel variant for clover-plus.
Mi Pad 4 Plus is very very similar to regular Mi Pad 4,
but it has different panel and different touchscreen.
Therefore it uses different device tree in kernel - and it
is the only changed line in deviceinfo file.

Signed-off-by: Alexey Minnekhanov <alexeymin@postmarketos.org>

[ci:skip-build]: Already built successfully in CI.