device-motorola-titan: remove Weston configuration (MR 4713)
The fbdev backend was removed with Weston 10 which has been out for a
good while now. As such, it doesn't matter whether Xwayland works or not
since Weston won't work at all --- the DRM backend, which is what would
get used instead of fbdev with this removed, won't work on downstream
kernels like the one used by this device.

[ci:ignore-count]
[ci:skip-build]: Already built successfully in CI.