device-xiaomi-surya: modules-initfs: Add ts+panel and drop gpi+spi
Add a touchscreen (nt36xxx-spi) and panel (panel-huaxing-nt36672c) to
modules-initfs because touchscreen was switched to a module in the last
release (which broke FDE for one release) and panel in this release.

Gpi and spi were enabled as build-in in the kernel (used on all sm7150
devices). So drop these module names from here.

Signed-off-by: Danila Tikhonov <danila@jiaxyga.com>