soc-qcom-sm7150: Implement accel libssc_180_degrees mount matrix
Introduce new udev rules tailored for Xiaomi devices (surya/davinci)
and potentially applicable to a broader range of sm7150 devices.
These rules are added to the SoC package and will be set as dependencies
in device packages for nonfree_firmware in subsequent commits.

Also use the -t argument to install other configs and udev rules. This
change does not introduce functional changes, thus it does not
necessitate a separate commit.

Co-authored-by: Jens Reidel <adrian@travitia.xyz>
Signed-off-by: Jens Reidel <adrian@travitia.xyz>
Signed-off-by: Danila Tikhonov <danila@jiaxyga.com>